requirements-dev.txt
-r requirements.txt

# Testing

pytest
pytest-django

# Optional: factories and helpers for tests

factory-boy
pytest-cov

# Lint / formatting (optional)

black
isort
flake8
